use java 8 date-time api

+++ src/main/kotlin/love/sola/spigot/dailysign/sql/UserInfo.kt
@@ -5,7 +5,7 @@ import java.util.*
 data class UserInfo(
     val player_id: UUID,
     val playerName: String,
-    var signCount: Int,
-    var continuousSignCount: Int,
-    var highestContinuous: Int
+    var signCount: Long,
+    var continuousSignCount: Long,
+    var highestContinuous: Long
 )
